% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data.R
\name{add_annotations_bio_cln}
\alias{add_annotations_bio_cln}
\title{Add columns}
\usage{
add_annotations_bio_cln(
  df,
  df_bio = NULL,
  df_cln = NULL,
  col_on_bio = NULL,
  col_on_cln = NULL,
  cols_bio = NULL,
  cols_cln = NULL
)
}
\arguments{
\item{df}{the dataframe to be annotated}

\item{df_bio}{the dataframe of biospecimen annotations}

\item{df_cln}{the dataframe of subjects annotations}

\item{col_on_bio}{column in the dataframe to be used to connect to df_bio}

\item{col_on_cln}{column in the dataframe to be used to connect to df_cln}

\item{cols_bio}{columns from the df_bio dataframe to be added to df}

\item{cols_cln}{columns from the df_cln dataframe to be added to df}
}
\description{
Add annotations from bio and cln files.
}
\author{
Yoann Pradat
}
