% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tests.R
\name{add_pvals_tables}
\alias{add_pvals_tables}
\title{Perform multiple testing correction from a set of pvalues}
\usage{
add_pvals_tables(
  dfs_plot,
  cohorts_a,
  cohorts_b,
  test,
  col_evt,
  in_plot_evt,
  tt_keep,
  n_cores,
  suffix = ""
)
}
\arguments{
\item{dfs_pvals}{a list of dataframes. Names must contain the values in \code{cohorts}.}

\item{cohorts}{A character vector.}

\item{method}{The multiple testing correction method.}
}
\value{
a list of dataframes. Names and format are identical to that of \code{dfs_pvals}.
}
\description{
From a list of multiple tables of pvalues, perform one multiple testing comparison correction considering all tests of
all tables at the same time.  Corrected pvalues (qvalues) are returned in the same format as the input list of
pvalue tables.
}
\author{
Yoann Pradat
}
