% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rna.R
\name{choose_columns}
\alias{choose_columns}
\title{Choose which columns should be loaded.}
\usage{
choose_columns(df_summary, study, metric, identifiers, identifiers_name)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose one of.
\itemize{
\item "met500": Trim Galore > Kallisto > Tximport. See \url{https://github.com/gustaveroussy/MetaPRISM}
\item "prism": Trim Galore > Kallisto > Tximport. See \url{https://github.com/gustaveroussy/MetaPRISM}
\item "tcga": From \url{https://stanfordmedicine.app.box.com/s/lu703xuaulfz02vgd2lunxnvt4mfvo3q}
\item "tcga_6_samples_prism": PRISM pipeline on 6 TCGA FASTQs.
\item "tcga_6_samples_gao": Gao pipeline on 6 TCGA FASTQs.
}}

\item{metric}{string or vector of strings. If not specified, every metric is loaded.
\itemize{
\item "TPM" to load "abundance.kallisto"
\item "counts" to load "counts.kallisto"
\item "length" to load "length.kallisto"
}}

\item{identifiers}{(optional) If not NULL, return RNA data only for the identifiers values. Used only if
\code{identifiers_name} is not NULL.}

\item{identifiers_name}{(optional) Field name of the identifiers.}

\item{df_meta}{a data.frame with summary data, loaded from \code{\link{load_summary_rna_gex}}}
}
\value{
a data.frame that contains part of the rows of \code{df_meta}
}
\description{
Based on a dataframe with summary data that links column names of the raw rna tables to subject and samples
identifiers, this function subsets the summary data on the exact set ofcolumns that will be loaded for the parameters
specified.
}
\author{
Yoann Pradat
}
\keyword{internal}
