% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tests.R
\name{compute_pvals_table}
\alias{compute_pvals_table}
\title{Compute pvalues of multiple 2x2 tables.}
\usage{
compute_pvals_table(
  dfs,
  cohort_a,
  cohort_b,
  row_col,
  row_names,
  col_names,
  test = "fisher",
  n_cores = 1
)
}
\arguments{
\item{dfs}{a 2-level list of data.frames. \code{cohort_a} and \code{cohort_b} must be 1st level names while 2nd-level
names must contain 'count' (the count matrix) and 'count_col' (the vector of columns margins).}

\item{cohort_a}{Name of the cohort A.}

\item{cohort_b}{Name of the cohort B.}

\item{row_col}{Name of the column to be used as rownames.}

\item{row_names}{Values of \code{row_col} to run tests on.}

\item{col_names}{Subcohort names of each of cohort A and B to run tests on.}

\item{test}{Name of the statistical test employed. Use "fisher" for Fisher-Booschloo, and "cmh" for CMH stratified
fisher test.}

\item{n_cores}{Number of cores to be used for parallel computations.}
}
\description{
From a list of 2 data.frames and their margins, build 2 tables with identical rows and columns and run Fisher tests
on each of the 2x2 contigency table built from each pair of cells of both tables.
}
\author{
Yoann Pradat
}
