% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load.R
\name{delineate_data}
\alias{delineate_data}
\title{Delineate treatments from compact format to more user-friendly format}
\usage{
delineate_data(
  df,
  col_name,
  format,
  separate_rows = T,
  prefix = "",
  suffix = "",
  delineate_dates = F,
  convert_dates_to_datetime = T,
  sep = ";",
  name = "Name"
)
}
\arguments{
\item{df}{a data.frame that must contain the column \code{col_name}.}

\item{col_name}{Name of the column to be delineated.}

\item{format}{Character vector of delineated column names corresponding the format of the compressed data.}

\item{separate_rows}{(optional) Should data be first unnested on the "|" separator"?}

\item{prefix}{(optional) Prefix to the new columns that contain the delineated data.}

\item{suffix}{(optional) Suffix to the new columns that contain the delineated data.}

\item{delineate_dates}{(optional) Used only if "Dates" is in \code{format}. Split "Dates" into "Date_Min" and
"Date_Max".}

\item{convert_dates_to_datetime}{(optional) Used only if delineate_dates=True. Should dates be converted to datetime
type?}

\item{sep}{(optional) The delimiter of the compressed data.}

\item{name}{(optional) Name given to the column containing the value before the brackets. Prefix and suffixes
specified will be added to this name.}
}
\value{
a data.frame with possibly more rows (if \code{separate_rows} is set to \code{TRUE}) and more columns
(as many as names in the list \code{format}) built from \code{df}.
}
\description{
Compressed data must consist of a name with additional data enclosed in brackets and separated by a common delimiter.
}
