% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{get_filepath_cln_curated}
\alias{get_filepath_cln_curated}
\title{Filepath to curated subject data files}
\usage{
get_filepath_cln_curated(study, mode)
}
\arguments{
\item{study}{Name of the study. Choose 'met500', 'prism' or 'tcga'.}

\item{mode}{Choose the file. 'all' for data about all subjects or 'in_design' for data about subjects in the design
only.}
}
\description{
Filepath to curated subject data files
}
\keyword{internal}
