% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{get_filepath_dsg_curated}
\alias{get_filepath_dsg_curated}
\title{Filepath to curated design data table}
\usage{
get_filepath_dsg_curated(study)
}
\arguments{
\item{study}{Name of the study. Only 'prism' is supported.}
}
\description{
Filepath to curated design data table
}
\keyword{internal}
