% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{get_filepath_ids_curated}
\alias{get_filepath_ids_curated}
\title{Filepath to curated id files}
\usage{
get_filepath_ids_curated(study)
}
\arguments{
\item{study}{Name of the study. Choose 'met500', 'prism' or 'tcga'.}
}
\description{
Filepath to curated id files
}
\keyword{internal}
