% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{get_filepath_res}
\alias{get_filepath_res}
\title{Filepath to a resource file from a database.}
\usage{
get_filepath_res(database, name)
}
\arguments{
\item{database}{Name of the database.}

\item{name}{Name of the resource.}
}
\description{
Filepath to a resource file from a database.
}
\keyword{internal}
