% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{get_filepath_summary_rna_fus}
\alias{get_filepath_summary_rna_fus}
\title{Filepath to summary rna-seq gene fusion tables}
\usage{
get_filepath_summary_rna_fus(study, mode)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose 'met500', 'prism' or 'tcga'.}

\item{mode}{The mode determines the caller and/or the filtering level.
\itemize{
\item "met500": arriba, ericscript, pizzly, starfusion, oncokb_civic.
\item "prism": arriba, ericscript, pizzly, starfusion, oncokb_civic.
\item "tcga": aggregated, oncokb_civic.
}}
}
\description{
Filepath to summary rna-seq gene fusion tables
}
\keyword{internal}
