% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepaths.R
\name{get_filepath_summary_wes_mut}
\alias{get_filepath_summary_wes_mut}
\title{Filepath to summary mutation tables}
\usage{
get_filepath_summary_wes_mut(study, mode)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose one of "met500", "prism" or "tcga".}

\item{mode}{(optional) Choose which MAF file to load.}
}
\description{
Filepath to summary mutation tables
}
\keyword{internal}
