% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tests.R
\name{get_pvals_cmh}
\alias{get_pvals_cmh}
\title{Compute pvalues of multiple stratified Fisher tests.}
\usage{
get_pvals_cmh(df_count_a, df_count_b, margins_a, margins_b, n_cores = 1)
}
\arguments{
\item{df_count_a}{a \code{data.frame}}

\item{df_count_b}{a \code{data.frame}}

\item{margins_a}{a numeric vector defining the sample size of each column of \code{df_count_a}}

\item{margins_b}{a numeric vector defining the sample size of each column of \code{df_count_b}}

\item{n_cores}{number of cores to be used for parallel computations.}
}
\value{
a \code{data.frame} with the same row names and column names as \code{df_count_a} and \code{df_count_b}.
}
\description{
From 2 matrices of event counts in variables and fixed samples, compute a stratified Fisher test
for each 2x2 count table induced by the each pair of rows in table A and table B.
}
\author{
Yoann Pradat
}
\keyword{internal}
