% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tests.R
\name{get_pvals_fisher}
\alias{get_pvals_fisher}
\title{Compute pvalues of multiple Fisher tests}
\usage{
get_pvals_fisher(
  df_count_a,
  df_count_b,
  margins_a,
  margins_b,
  conf_int = F,
  conf_lvl = 0.95,
  progress = F,
  n_cores = 1
)
}
\arguments{
\item{df_count_a}{a \code{data.frame}.}

\item{df_count_b}{a \code{data.frame}.}

\item{margins_a}{a numeric vector defining the sample size of each column of \code{df_count_a}.}

\item{margins_b}{a numeric vector defining the sample size of each column of \code{df_count_b}.}

\item{conf_int}{if set to TRUE, the returned object is a list with names 'pvals', 'conf_int_l', 'conf_int_h'.}

\item{conf_lvl}{Use only if conf_int set to TRUE}

\item{progress}{set to TRUE to see progress bar.}

\item{n_cores}{number of cores to be used for parallel computations.}
}
\value{
a \code{data.frame} with the same row names and column names as \code{df_count_a} and \code{df_count_b}.
}
\description{
From 2 matrices of event counts in variables and fixed samples, compute a Fisher test for each 2x2 count table
induced by the each pair of cell in table A and table B.
}
\author{
Yoann Pradat
}
\keyword{internal}
