% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ids.R
\name{get_sample_id_from_germline}
\alias{get_sample_id_from_germline}
\title{Correct id of normal sample.}
\usage{
get_sample_id_from_germline(x)
}
\arguments{
\item{x}{sample id}
}
\description{
Correct id of normal sample.
}
\keyword{internal}
