% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ids.R
\name{get_sample_id_from_somatic}
\alias{get_sample_id_from_somatic}
\title{Correct id of tumor sample.}
\usage{
get_sample_id_from_somatic(x)
}
\arguments{
\item{x}{sample id}
}
\description{
Correct id of tumor sample.
}
\keyword{internal}
