% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bio.R
\name{load_bio}
\alias{load_bio}
\title{Load biospecimen files.}
\usage{
load_bio(
  study,
  identifiers = NULL,
  identifiers_name = NULL,
  mode = "in_design",
  ...
)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose "met500", "prism" or "tcga".}

\item{identifiers}{(optional) If not NULL, return clinical tables only for the individuals specified}

\item{identifiers_name}{(optional) The name of the column the \code{identifiers} values correspond to.}

\item{mode}{(optional) Choose "in_design" to load only data for samples in the design or 'all' to load data for all
samples of the cohort.}

\item{...}{Extra parameters passed to \code{\link{load_from_data}}}
}
\value{
a data.frame
}
\description{
This function loads biospecimen tables (one line is one sample id) for the specified study and optionally specified
identifiers. The \code{mode} parameters allow to choose which table should be loaded between 'in_design' and 'all'.
}
\author{
Yoann Pradat
}
