% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cln.R
\name{load_cln}
\alias{load_cln}
\title{Load clinical files.}
\usage{
load_cln(
  study,
  identifiers = NULL,
  identifiers_name = NULL,
  mode = "in_design",
  ...
)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose "met500", "prism" or "tcga".}

\item{identifiers}{(optional) If not NULL, return clinical tables only for the individuals specified.}

\item{identifiers_name}{(optional) The name of the column the \code{identifiers} values correspond to.}

\item{mode}{(optional) Either 'in_design' or 'all' for 'pancancer' annotations or 'brca' for per-cancer annotations.}

\item{...}{Extra parameters passed to \code{\link{load_from_data}}}
}
\value{
a data.frame
}
\description{
This function loads clinical tables (one line is one subject id) for the specified study and optionally specified
identifiers. The \code{mode} parameters allow to choose which table should be loaded between 'in_design' and 'all'.
}
\author{
Yoann Pradat
}
