% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_colors.R
\name{load_colors}
\alias{load_colors}
\title{Load lists of colors from Excel table}
\usage{
load_colors(sheet, as_tibble = F)
}
\arguments{
\item{sheet}{Name of the excel spreadsheet.}

\item{as_tibble}{(optional) Set to TRUE if you would like to have the list of colors returned in a \code{tibble}
dataframe. If FALSE, colors are returned as a named list.}
}
\value{
A named list or a \code{tibble} dataframe. Names are categories of some fields or custom names and values are
colors in hexadecimal character vectors.
}
\description{
Load one or multiple sheets from an excel workbook containing lists of colors for categories of different fields.
}
