% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dsg.R
\name{load_dsg}
\alias{load_dsg}
\title{Load the design table}
\usage{
load_dsg(study)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose "met500", "prism" or "tcga".}
}
\value{
a data.frame
}
\description{
Load the design table for the specified study. Only "prism" is supported.
}
\author{
Yoann Pradat
}
