% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load.R
\name{load_from_data}
\alias{load_from_data}
\title{Load a table located in the data folder}
\usage{
load_from_data(path, ...)
}
\arguments{
\item{path}{Path to the file.}

\item{...}{Extra parameters passed to \code{read_delim}.}
}
\description{
Load a table located in the data folder
}
