% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ids.R
\name{load_ids}
\alias{load_ids}
\title{Load identifiers files.}
\usage{
load_ids(study, ...)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose "met500", "prism" or "tcga".}

\item{...}{Extra parameters passed to \code{\link{load_from_data}}}
}
\value{
a data.frame
}
\description{
This function loads the table linking all subject, sample and idspsy identifiers..
}
\author{
Yoann Pradat
}
