% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_res.R
\name{load_resource}
\alias{load_resource}
\title{Load resources from different databases or used for different analyses.}
\usage{
load_resource(database, name, ...)
}
\arguments{
\item{database}{Name of the database to be load from. Supported values are
\itemize{
\item{civic}
\item{oncokb}
\item{cosmic}
\item{gencode}
\item{curated}
}}

\item{name}{Name of the resource}

\item{...}{Extra parameters passed to \code{\link{load_from_data}}}
}
\description{
Load resources from different databases or used for different analyses.
}
\author{
Yoann Pradat
}
