% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rna.R
\name{load_rna_fus}
\alias{load_rna_fus}
\title{Load fusions called from rna-seq data.}
\usage{
load_rna_fus(
  study,
  mode = NULL,
  identifiers = NULL,
  identifiers_name = NULL,
  ...
)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose 'met500', 'prism' or 'tcga'.}

\item{mode}{The mode determines the caller and/or the filtering level.
\itemize{
\item "met500": arriba, ericscript, pizzly, starfusion, oncokb_civic.
\item "prism": arriba, ericscript, pizzly, starfusion, oncokb_civic.
\item "tcga": aggregated, oncokb_civic.
}}

\item{identifiers}{(optional) If not NULL, return RNA data only for the identifiers values. Used only if
\code{identifiers_name} is not NULL.}

\item{identifiers_name}{(optional) Field name of the identifiers.}

\item{...}{Extra parameters passed to \code{\link{load_from_data}}}
}
\value{
a tibble data.frame
}
\description{
This function loads RNA gene fusion tables for the specified cohort or subcohort.
}
\author{
Yoann Pradat
}
