% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rna.R
\name{load_rna_gex}
\alias{load_rna_gex}
\title{Load raw gene expression data files.}
\usage{
load_rna_gex(
  study,
  level = c("genes", "transcripts"),
  metric = c("counts", "TPM", "length"),
  test_mode = F,
  identifiers = NULL,
  identifiers_name = NULL,
  ...
)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose one of.
\itemize{
\item "met500": Trim Galore > Kallisto > Tximport. See \url{https://github.com/gustaveroussy/MetaPRISM}
\item "prism": Trim Galore > Kallisto > Tximport. See \url{https://github.com/gustaveroussy/MetaPRISM}
\item "tcga": From \url{https://stanfordmedicine.app.box.com/s/lu703xuaulfz02vgd2lunxnvt4mfvo3q}
\item "tcga_6_samples_prism": PRISM pipeline on 6 TCGA FASTQs.
\item "tcga_6_samples_gao": Gao pipeline on 6 TCGA FASTQs.
}}

\item{level}{Level at which counts are aggregated. You may choose "genes" or "transcripts".}

\item{metric}{string or vector of strings. If not specified, every metric is loaded.
\itemize{
\item "TPM" to load "abundance.kallisto"
\item "counts" to load "counts.kallisto"
\item "length" to load "length.kallisto"
}}

\item{test_mode}{(optional) Set to True to use submatrices generated with \code{generate_rna_test} from
\code{_util_rna} module.}

\item{identifiers}{(optional) If not NULL, return RNA data only for the identifiers values. Used only if
\code{identifiers_name} is not NULL.}

\item{identifiers_name}{(optional) Field name of the identifiers.}

\item{...}{Extra parameters passed to \code{\link{load_from_data}}}
}
\value{
a tibble data.frame
}
\description{
This function loads RNA gene expression tables for the specified cohort or subcohort.
}
\author{
Yoann Pradat
}
