% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_summary.R
\name{load_summary_rna_fus}
\alias{load_summary_rna_fus}
\title{Load summary of rna-seq gene fusion tables}
\usage{
load_summary_rna_fus(study, mode = NULL)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose 'met500', 'prism' or 'tcga'.}

\item{mode}{The mode determines the caller and/or the filtering level.
\itemize{
\item "met500": arriba, ericscript, pizzly, starfusion, oncokb_civic.
\item "prism": arriba, ericscript, pizzly, starfusion, oncokb_civic.
\item "tcga": aggregated, oncokb_civic.
}}
}
\value{
a data.frame with summary
}
\description{
The summary table links Sample_Id to Subject_Id. This allows easier connection
to other data tables using either of Sample_Id or Subject_Id.
}
