% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_summary.R
\name{load_summary_rna_gex}
\alias{load_summary_rna_gex}
\title{Load summary of RNA tables}
\usage{
load_summary_rna_gex(study, level, metric, test_mode = F)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose one of.
\itemize{
\item "met500": Trim Galore > Kallisto > Tximport. See \url{https://github.com/gustaveroussy/MetaPRISM}
\item "prism": Trim Galore > Kallisto > Tximport. See \url{https://github.com/gustaveroussy/MetaPRISM}
\item "tcga": From \url{https://stanfordmedicine.app.box.com/s/lu703xuaulfz02vgd2lunxnvt4mfvo3q}
\item "tcga_6_samples_prism": PRISM pipeline on 6 TCGA FASTQs.
\item "tcga_6_samples_gao": Gao pipeline on 6 TCGA FASTQs.
}}

\item{level}{Level at which counts are aggregated. You may choose "genes" or "transcripts".}

\item{metric}{string or vector of strings. If not specified, every metric is loaded.
\itemize{
\item "TPM" to load "abundance.kallisto"
\item "counts" to load "counts.kallisto"
\item "length" to load "length.kallisto"
}}

\item{test_mode}{(optional) Set to True to use submatrices generated with \code{generate_rna_test} from
\code{_util_rna} module.}
}
\value{
a data.frame with summary
}
\description{
The summary table links original names in the RNA table to Sample_Id and Subject_Id. This allows easier connection
to other data tables using either of Sample_Id or Subject_Id.
}
