% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load.R
\name{load_table}
\alias{load_table}
\title{Load a table.}
\usage{
load_table(path, header_prefix = NULL, ...)
}
\arguments{
\item{path}{Path to the file.}

\item{header_prefix}{If not NULL, skip header rows that will be identified as first lines starting with this prefix.}

\item{...}{Extra parameters passed to \code{read_delim}.}
}
\description{
Load a table.
}
