% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wes.R
\name{load_wes_mut}
\alias{load_wes_mut}
\title{Load mutation tables.}
\usage{
load_wes_mut(
  study,
  identifiers = NULL,
  identifiers_name = NULL,
  mode = "somatic_maf",
  ...
)
}
\arguments{
\item{study}{Name of the cohort used in the naming of the files. Choose one of "met500", "prism" or "tcga".}

\item{identifiers}{(optional) If not NULL, return clinical tables only for the individuals specified}

\item{identifiers_name}{(optional) The name of the column the \code{identifiers} values correspond to.}

\item{mode}{(optional) Choose which MAF file to load.}

\item{...}{Extra parameters passed to \code{\link{load_from_data}}}
}
\value{
a \code{data.frame}
}
\description{
Load mutation tables for the specified study and optionally specified identifiers. This function is an R
reimplementation of the function \code{load_wes_mut} from the pyprism package.
}
\details{
\itemize{
\item If "somatic_maf", load data/\link{study}/wes/somatic_maf/somatic_calls.maf.gz
\item If "somatic_filters", load data/\link{study}/wes/somatic_maf/somatic_calls_filters.tsv.gz
\item If "somatic_oncokb", load data/\link{study}/wes/somatic_maf/somatic_calls_oncokb.maf.gz
\item If "somatic_civic", load data/\link{study}/wes/somatic_maf/somatic_calls_civic.maf.gz
}
}
\author{
Yoann Pradat
}
