% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_on_rows}
\alias{merge_on_rows}
\title{Merge two dataframes on their row names.}
\usage{
merge_on_rows(df_x, df_y, how_cols = "inner")
}
\arguments{
\item{df_x}{a data.frame}

\item{df_y}{a data.frame}

\item{how_cols}{a character vector specifying how the merge should proceed
\enumerate{
\item 'inner' keep only the columns common to both \code{df_x} and \code{df_y}
\item 'x' keep only the columns of \code{df_x}
\item 'y' keep only the columns of \code{df_y}
\item 'outer' keep all the columns of \code{df_x} and \code{df_y}
}

This function raises an error if a row entry (identified by its row name) is present in both \code{df_x} and
\code{df_y} and has divergent values on a column that is present in both data frames.}
}
\value{
a data.frame
}
\description{
This function merges two dataframes on their row names and provides the user with the choice on which columns should
be kept.
}
\author{
Yoann Pradat
}
\keyword{export}
