% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tests.R
\name{prepare_count_table}
\alias{prepare_count_table}
\title{Add missing rows and columns.}
\usage{
prepare_count_table(df, row_names, col_names, row_col)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{row_names}{a character vector}

\item{col_names}{a character vector}

\item{row_col}{column to be used as rownames for \code{df}.}
}
\value{
a \code{data.frame} with missing rows and columns added.
}
\description{
From a table and a list of row names and column names, add rows or columns with all 0s if they are not already in the
table. After missing rows or columns were added, both rows and columns are sorted.
}
\author{
Yoann Pradat
}
\keyword{internal}
