% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tests.R
\name{prepare_margins}
\alias{prepare_margins}
\title{Add missing row names in margins.}
\usage{
prepare_margins(df, row_names)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{row_names}{}
}
\description{
After missing rows were added, rows are sorted.
}
\author{
Yoann Pradat
}
\keyword{internal}
