% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data.R
\name{preprocess_wes_mut}
\alias{preprocess_wes_mut}
\title{Preprocess mutations table}
\usage{
preprocess_wes_mut(
  df_mut,
  cohort,
  cols_bio = c(),
  cols_cln = c(),
  select_pairs = FALSE,
  selection_mut = "all",
  verbose = TRUE
)
}
\arguments{
\item{df_mut}{data.frame of mutations}

\item{cohort}{name of the cohort}

\item{cols_bio}{(optional) list ist of bio attributes}

\item{cols_cln}{(optional) list of cln attributes.}

\item{select_pairs}{(optional) If set to TRUE, 1 pair tumor/normal is selected for each subject.}

\item{selection_mut}{(optional)
Here are the available modes
\enumerate{
\item{'all'}{no selection of variants}
\item{'annotated'}{no selection of variants}
\item{'non_synonymous'}{select variants for which the value of 'Variant_Classification' is one of
* Frame_Shift_Del
* Frame_Shift_Ins,
* Splice_Site
* Translation_Start_Site
* Nonsense_Mutation
* Nonstop_Mutation
* In_Frame_Del
* In_Frame_Ins
* Missense_Mutation
* Start_Codon_Del
* Start_Codon_SNP
* Stop_Codon_Del
* Stop_Codon_Ins}
\item{'truncating'}{select variants for which the value of 'Variant_Classification' is one of
* Frame_Shift_Del
* Frame_Shift_Ins,
* Splice_Site
* Nonsense_Mutation}
}}

\item{verbose}{(optional) Should info messages be printed?}
}
\value{
A dataframe of mutations with possibly additional attributes and possibly a reduced number of lines.
}
\description{
Add bio and/or cln attributes to the mutations table and perform a selection of one pair tumor/normal per subject
if requested.
}
