% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progress}
\alias{progress}
\title{A progress bar function to run within a \code{for} loop}
\usage{
progress(i, n)
}
\arguments{
\item{i}{A numeric value indicating the current iteration of the containing for loop}

\item{n}{A numeric value indicating the total number of iterations of the containing for loop}
}
\description{
This function should be run within a \code{for} loop, it produces a progress bar to indicate how many iterations have
passed
}
\author{
Joseph Crispell see repo \url{https://github.com/JosephCrispell/basicPlotteR}
}
\keyword{internal}
