% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_load.R
\name{read_header}
\alias{read_header}
\title{Load header of a table.}
\usage{
read_header(path, prefix)
}
\arguments{
\item{path}{Path to the file.}

\item{prefix}{Prefix for identifying header}
}
\description{
Load header of a table.
}
\keyword{internal}
