% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprism.R
\docType{package}
\name{rprism}
\alias{rprism}
\title{rprism: An R package accompanying the analysis of the META-PRISM cohort}
\description{
The rprism package provides handy functions for curating data, loading row and curated data
and for performing recurrent operatons on the data. It also provides util functions for running statistical tests or
for setting the working directory to specific locations for instance.
}
