% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data.R
\name{select_tumor_types}
\alias{select_tumor_types}
\title{Function for selecting tumor types.}
\usage{
select_tumor_types(dfs, col_tt, tt_min_size, tt_drop)
}
\arguments{
\item{dfs}{A named list of \code{data.frame}. They must contain columns \code{col_tt}.}

\item{tt_min_size}{Numeric value.}

\item{tt_drop}{Values that should be discarded regarless of the size.}
}
\value{
a list containing a named list of dataframes and a named list of character vectors, each with the sames names
as \code{dfs}.
}
\description{
Using the column \code{col_tt}, this function subsets each input dataframe to the values of \code{col_tt} with at
least \code{tt_min_size} entries.
}
\author{
Yoann Pradat
}
