% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd.R
\name{setwd_to_results}
\alias{setwd_to_results}
\title{Set the working directory to the results folder of the project.}
\usage{
setwd_to_results()
}
\value{
current_wd name of the working directory before change
}
\description{
Set the working directory to the results folder of the project.
}
\author{
Yoann Pradat
}
