% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data.R
\name{split_targeted_therapy_targets}
\alias{split_targeted_therapy_targets}
\title{Split targeted therapy classes with multiple targets}
\usage{
split_targeted_therapy_targets(x, sep = "|")
}
\arguments{
\item{x}{a string containing entries (e.g drug names) separated by sep}

\item{sep}{the separator, default is "|"}
}
\value{
a string
}
\description{
Split multiple-targets therapies into multiple single-target therapies.
}
